refactor: little fixes
