+++ package.json
@@ -2,7 +2,7 @@
   "name": "exact-hunter",
   "version": "0.1.0",
   "private": true,
-  "packageManager": "pnpm@10.32.1",
+  "packageManager": "pnpm@10.33.0",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -28,7 +28,7 @@
     "clsx": "^2.1.1",
     "immer": "^11.1.4",
     "lucide-react": "^0.574.0",
-    "next": "16.1.6",
+    "next": "^16.2.1",
     "radix-ui": "^1.4.3",
     "react": "19.2.3",
     "react-dom": "19.2.3",
@@ -45,9 +45,9 @@
     "@dreamsicle.io/stylelint-config-tailwindcss": "^1.2.2",
     "@jonasgeiler/tsc-files": "^2.3.1",
     "@storybook/addon-a11y": "^10.2.18",
-    "@storybook/addon-docs": "^10.2.18",
+    "@storybook/addon-docs": "^10.3.3",
     "@storybook/addon-vitest": "^10.2.18",
-    "@storybook/nextjs-vite": "^10.2.18",
+    "@storybook/nextjs-vite": "^10.3.3",
     "@svgr/plugin-svgo": "^8.1.0",
     "@svgr/webpack": "^8.1.0",
     "@tailwindcss/postcss": "^4",
@@ -56,8 +56,8 @@
     "@types/react-dom": "^19",
     "@vitest/browser-playwright": "^4.0.18",
     "@vitest/coverage-v8": "^4.0.18",
-    "eslint": "^9",
-    "eslint-config-next": "16.1.6",
+    "eslint": "^9.39.4",
+    "eslint-config-next": "16.2.1",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-fsd-lint": "^1.0.10",
     "eslint-plugin-import": "^2.32.0",
@@ -66,7 +66,7 @@
     "eslint-plugin-storybook": "^10.2.18",
     "husky": "^9.1.7",
     "jsdom": "^29.0.1",
-    "lint-staged": "^16.2.7",
+    "lint-staged": "^16.4.0",
     "msw": "^2.12.10",
     "playwright": "^1.58.2",
     "prettier": "^3.8.1",
