+++ src/app/providers/QueryProvider.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import { useState } from 'react';
+
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 const QueryProvider = ({ children }: { children: React.ReactNode }) => {
