+++ src/entities/session/model/api/session.service.ts
@@ -1,4 +1,4 @@
-import { RefreshResponse } from '../types/RefreshResp';
+import { RefreshResponse } from '@session/model/types/RefreshResp';
 
 import { mutation } from '@/shared/api/api';
