+++ src/entities/session/model/hooks/useLogoutMutation.ts
@@ -1,7 +1,7 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { logoutFn } from '@session/model/api/session.service';
+import { useAuthStore } from '@session/model/store/auth.store';
 
-import { logoutFn } from '../api/session.service';
-import { useAuthStore } from '../store/auth.store';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 export const useLogoutMutation = () => {
   const queryClient = useQueryClient();
