+++ src/entities/session/model/hooks/useSessionRefreshQuery.ts
@@ -1,8 +1,9 @@
 import { useCallback, useRef } from 'react';
-import { useMutation } from '@tanstack/react-query';
 
-import { refreshAuthTokenFn } from '../api/session.service';
-import { useAuthStore } from '../store/auth.store';
+import { refreshAuthTokenFn } from '@session/model/api/session.service';
+import { useAuthStore } from '@session/model/store/auth.store';
+
+import { useMutation } from '@tanstack/react-query';
 
 const REFRESH_URL = '/auth/refresh';
