+++ src/features/auth/model/hooks/useAttachAuthToken.ts
@@ -1,6 +1,7 @@
 'use client';
 
 import { useMemo } from 'react';
+
 import type { InternalAxiosRequestConfig } from 'axios';
 
 import { useAuthStore } from '@/entities/session';
