+++ src/shared/lib/index.ts
@@ -1,3 +1,4 @@
 import * as ErrorGuards from './guards/error-guards';
 
 export { ErrorGuards as EG };
+export { getFieldError } from './helpers/getFieldError';
