+++ src/shared/ui/label/Label.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
 
-import { Input } from '../input/Input';
+import { Input } from '../input';
 
 import { Label } from './Label';
